Rollback and segregation of business logic from REST implementation
Rollback implemented for:
1. Create volume - different variations
2. Create clone
3. Create snapshot

Segregation:
VolumePlugin class was getting cluttered with business logic. Moved that out to a different class VolumeManager.
