+++ hpedockerplugin/hpe_storage_api.py
@@ -17,28 +17,22 @@ An HTTP API implementing the Docker Volumes Plugin API.
 
 See https://github.com/docker/docker/tree/master/docs/extend for details.
 """
-from os_brick.initiator import connector
-from oslo_utils import netutils
-import uuid
-from i18n import _, _LE, _LI, _LW
-import exception
+import json
 import six
 
-import json
-
+from os_brick.initiator import connector
+from oslo_log import log as logging
+from oslo_utils import netutils
 from twisted.python.filepath import FilePath
 
+import exception
 import fileutil
-
+from i18n import _, _LE, _LI, _LW
 from klein import Klein
 from hpe import volume
-from hpe import utils
 from oslo_utils import importutils
 import etcdutil as util
-import synchronization
-
-from oslo_log import log as logging
-
+import volume_manager as mgr
 
 LOG = logging.getLogger(__name__)
 
@@ -100,6 +94,8 @@ class VolumePlugin(object):
         self.enforce_multipath = True
 
         self.connector = self._get_connector(protocol)
+        self._manager = mgr.VolumeManager(self.hpeplugin_driver,
+                                          self._etcd)
 
     def _get_connector(self, protocol):
         root_helper = 'sudo'
@@ -160,132 +156,11 @@ class VolumePlugin(object):
             # 1. We want to avoid regression in existing remove-volume
             # 2. In the future, if docker engine provides snapshot
             #    support, this code should have minimum impact
-            return self.volumedriver_remove_snapshot(volname, snapname)
+            return self._manager.remove_snapshot(volname, snapname)
         else:
             volname = tokens[0]
 
-        return self._volumedriver_remove(volname)
-
-    @synchronization.synchronized('{volname}')
-    def _volumedriver_remove(self, volname):
-        # Only 1 node in a multinode cluster can try to remove the volume.
-        # Grab lock for volume name. If lock is inuse, just return with no
-        # error.
-        # Expand lock code inline as function based lock causes
-        # unexpected behavior
-        vol = self._etcd.get_vol_byname(volname)
-        if vol is None:
-            # Just log an error, but don't fail the docker rm command
-            msg = (_LE('Volume remove name not found %s'), volname)
-            LOG.error(msg)
-            return json.dumps({u"Err": ''})
-
-        try:
-            if vol['snapshots']:
-                msg = (_LE('Err: Volume %s has one or more child '
-                           'snapshots - volume cannot be deleted!'
-                           % volname))
-                LOG.error(msg)
-                response = json.dumps({u"Err": msg})
-                return response
-            else:
-                self.hpeplugin_driver.delete_volume(vol)
-                LOG.info(_LI('volume: %(name)s,' 'was successfully deleted'),
-                         {'name': volname})
-        except Exception as ex:
-            msg = (_LE('Err: Failed to remove volume %s, error is %s'),
-                   volname, six.text_type(ex))
-            LOG.error(msg)
-            return json.dumps({u"Err": six.text_type(ex)})
-
-        try:
-            self._etcd.delete_vol(vol)
-        except KeyError:
-            msg = (_LW('Warning: Failed to delete volume key: %s from '
-                       'etcd due to KeyError'), volname)
-            LOG.warning(msg)
-            pass
-        return json.dumps({u"Err": ''})
-
-    def _get_snapshot_by_name(self, snapshots, snapname):
-        idx = 0
-        for s in snapshots:
-            if s['name'] == snapname:
-                return s, idx
-            idx = idx + 1
-        return None, None
-
-    @synchronization.synchronized('{volname}')
-    def volumedriver_remove_snapshot(self, volname, snapname):
-        LOG.info("volumedriver_remove_snapshot - getting volume %s"
-                 % volname)
-
-        vol = self._etcd.get_vol_byname(volname)
-        if vol is None:
-            # Just log an error, but don't fail the docker rm command
-            msg = (_LE('snapshot remove - parent volume name not found '
-                       '%s'), volname)
-            LOG.error(msg)
-            return json.dumps({u"Err": msg})
-
-        if snapname:
-            snapshots = vol['snapshots']
-            LOG.info("Getting snapshot by name: %s" % snapname)
-            snapshot, idx = self._get_snapshot_by_name(snapshots,
-                                                       snapname)
-
-            if snapshot:
-                LOG.info("Found snapshot by name: %s" % snapname)
-                # Does the snapshot have child snapshot(s)?
-                for ss in snapshots:
-                    LOG.info("Checking if snapshot has children: %s"
-                             % snapname)
-                    if ss['parent_id'] == snapshot['id']:
-                        msg = (_LE('snapshot %s/%s has one or more child '
-                                   'snapshots - it cannot be deleted!'
-                                   % (volname, snapname)))
-                        LOG.error(msg)
-                        response = json.dumps({u"Err": msg})
-                        return response
-                try:
-                    LOG.info("Deleting snapshot at backend: %s" % snapname)
-                    self.hpeplugin_driver.delete_volume(snapshot,
-                                                        is_snapshot=True)
-                except Exception as err:
-                    msg = (_LE('Failed to remove snapshot error is: %s'),
-                           six.text_type(err))
-                    LOG.error(msg)
-                    response = json.dumps({u"Err": six.text_type(err)})
-                    return response
-
-                LOG.info("Deleting snapshot in ETCD - %s" % snapname)
-                # Remove snapshot entry from list and save it back to
-                # ETCD DB
-                del snapshots[idx]
-                try:
-                    LOG.info("Updating volume in ETCD after snapshot "
-                             "removal - vol-name: %s" % volname)
-                    # For now just track volume to uuid mapping internally
-                    # TODO: Save volume name and uuid mapping in etcd as
-                    # well. This will make get_vol_byname more efficient
-                    self._etcd.update_vol(vol['id'],
-                                          'snapshots',
-                                          snapshots)
-                    LOG.info('snapshot: %(name)s was successfully '
-                             'removed', {'name': snapname})
-                    response = json.dumps({u"Err": ''})
-                    return response
-                except Exception as ex:
-                    msg = (_('remove snapshot from etcd failed, error is:'
-                             ' %s'), six.text_type(ex))
-                    LOG.error(msg)
-                    response = json.dumps({u"Err": six.text_type(ex)})
-                    return response
-            else:
-                msg = (_LE('snapshot %s does not exist!' % snapname))
-                LOG.error(msg)
-                response = json.dumps({u"Err": msg})
-                return response
+        return self._manager.remove_volume(volname)
 
     @app.route("/VolumeDriver.Unmount", methods=["POST"])
     def volumedriver_unmount(self, name):
@@ -469,46 +344,9 @@ class VolumePlugin(object):
                 LOG.error(msg)
                 return json.dumps({u"Err": six.text_type(msg)})
 
-        response = self._volumedriver_create(volname, vol_size,
-                                             vol_prov, vol_flash,
-                                             compression_val, vol_qos)
-        return response
-
-    @synchronization.synchronized('{volname}')
-    def _volumedriver_create(self, volname, vol_size, vol_prov,
-                             vol_flash, compression_val, vol_qos):
-        LOG.debug('In _volumedriver_create')
-
-        # NOTE: Since Docker passes user supplied names and not a unique
-        # uuid, we can't allow duplicate volume names to exist
-        vol = self._etcd.get_vol_byname(volname)
-        if vol is not None:
-            return json.dumps({u"Err": ''})
-
-        vol = volume.createvol(volname, vol_size, vol_prov,
-                               vol_flash, compression_val, vol_qos)
-        try:
-            self.hpeplugin_driver.create_volume(vol)
-        except Exception as ex:
-            msg = (_('Create volume failed with error: %s'), six.text_type(ex))
-            LOG.exception(msg)
-            return json.dumps({u"Err": six.text_type(ex)})
-
-        response = json.dumps({u"Err": ''})
-        try:
-            # For now just track volume to uuid mapping internally
-            # TODO: Save volume name and uuid mapping in etcd as well
-            # This will make get_vol_byname more efficient
-            self._etcd.save_vol(vol)
-            LOG.debug('volume: %(name)s was successfully saved to etcd',
-                      {'name': volname})
-        except exception.HPEPluginSaveFailed as ex:
-            # TODO: Cleanup backend volume and if needed VVS too
-            msg = 'Save volume to etcd failed, error is: %s'\
-                  % six.text_type(ex)
-            LOG.exception(msg)
-            response = json.dumps({u"Err": six.text_type(_(ex))})
-        return response
+        return self._manager.create_volume(volname, vol_size,
+                                           vol_prov, vol_flash,
+                                           compression_val, vol_qos)
 
     def volumedriver_clone_volume(self, name, opts=None):
         # Repeating the validation here in anticipation that when
@@ -520,65 +358,14 @@ class VolumePlugin(object):
             LOG.error(msg)
             raise exception.HPEPluginCreateException(reason=msg)
 
-        src_vol_name = str(contents['Opts']['cloneOf'])
-        clone_name = contents['Name']
-
-        return self._volumedriver_clone_volume(
-            src_vol_name, clone_name, contents)
-
-    @synchronization.synchronized('{src_vol_name}')
-    def _volumedriver_clone_volume(self, src_vol_name, clone_name,
-                                   contents):
-        # Check if volume is present in database
-        src_vol = self._etcd.get_vol_byname(src_vol_name)
-        if src_vol is None:
-            msg = 'source volume: %s does not exist' % src_vol_name
-            LOG.debug(msg)
-            response = json.dumps({u"Err": msg})
-            return response
-
+        size = None
         if ('Opts' in contents and contents['Opts'] and
                 'size' in contents['Opts']):
             size = int(contents['Opts']['size'])
-        else:
-            size = src_vol['size']
 
-        if size < src_vol['size']:
-            msg = 'clone volume size %s is less than source ' \
-                  'volume size %s' % (size, src_vol['size'])
-            LOG.debug(msg)
-            response = json.dumps({u"Err": msg})
-            return response
-        return self._clone_volume(clone_name, src_vol, size)
-
-    @synchronization.synchronized('{clone_name}')
-    def _clone_volume(self, clone_name, src_vol, size):
-        # Create clone volume specification
-        clone_vol = volume.createvol(clone_name, size,
-                                     src_vol['provisioning'],
-                                     src_vol['flash_cache'],
-                                     src_vol['compression'],
-                                     src_vol['qos_name'])
-        try:
-            self.hpeplugin_driver.create_cloned_volume(clone_vol, src_vol)
-
-            response = json.dumps({u"Err": ''})
-            # For now just track volume to uuid mapping internally
-            # TODO: Save volume name and uuid mapping in etcd as well
-            # This will make get_vol_byname more efficient
-            self._etcd.save_vol(clone_vol)
-        except exception.HPEPluginEtcdException as ex:
-            # TODO: 3PAR clean up issue over here - clone got created
-            # in the backend but since it could not be saved in etcd db
-            # we are throwing an error saying operation failed.
-            # TODO: This needs to be fixed
-            response = json.dumps({u"Err": ex.message})
-        except Exception as ex:
-            msg = (_('clone volume failed, error is: %s'),
-                   six.text_type(ex))
-            LOG.error(msg)
-            response = json.dumps({u"Err": six.text_type(ex)})
-        return response
+        src_vol_name = str(contents['Opts']['cloneOf'])
+        clone_name = contents['Name']
+        return self._manager.clone_volume(src_vol_name, clone_name, size)
 
     def volumedriver_create_snapshot(self, name, opts=None):
         # Repeating the validation here in anticipation that when
@@ -619,75 +406,10 @@ class VolumePlugin(object):
         if 'Opts' in contents and contents['Opts'] and \
                 'retentionHours' in contents['Opts']:
             retention_hrs = int(contents['Opts']['retentionHours'])
-        return self._volumedriver_create_snapshot(src_vol_name,
-                                                  snapshot_name,
-                                                  expiration_hrs,
-                                                  retention_hrs)
-
-    @synchronization.synchronized('{src_vol_name}')
-    def _volumedriver_create_snapshot(self, src_vol_name, snapshot_name,
-                                      expiration_hrs, retention_hrs):
-        # Check if volume is present in database
-        vol = self._etcd.get_vol_byname(src_vol_name)
-        if vol is None:
-            msg = 'source volume: %s does not exist' % src_vol_name
-            LOG.debug(msg)
-            response = json.dumps({u"Err": msg})
-            return response
-
-        if vol['snapshots']:
-            ss_list = vol['snapshots']
-            for ss in ss_list:
-                if snapshot_name == ss['name']:
-                    msg = (_('Snapshot create failed. Error '
-                             'is: %(snap_name)s is already created. '
-                             'Please enter a new snapshot name.') %
-                           {'snap_name': snapshot_name})
-                    LOG.error(msg)
-                    return json.dumps({u"Err": six.text_type(msg)})
-
-        snapshot_id = str(uuid.uuid4())
-        snapshot = {'id': snapshot_id,
-                    'display_name': snapshot_name,
-                    'volume_id': vol['id'],
-                    'volume_name': src_vol_name,
-                    'expirationHours': expiration_hrs,
-                    'retentionHours': retention_hrs,
-                    'display_description': 'snapshot of volume %s'
-                                           % src_vol_name}
-        try:
-            self.hpeplugin_driver.create_snapshot(snapshot)
-        except Exception as ex:
-            msg = (_('create snapshot failed, error is: %s'),
-                   six.text_type(ex))
-            LOG.error(msg)
-            return json.dumps({u"Err": six.text_type(ex)})
-
-        response = json.dumps({u"Err": ''})
-        db_snapshot = {'name': snapshot_name,
-                       'id': snapshot_id,
-                       'parent_id': vol['id'],
-                       'expiration_hours': expiration_hrs,
-                       'retention_hours': retention_hrs}
-        vol['snapshots'].append(db_snapshot)
-        try:
-            # For now just track volume to uuid mapping internally
-            # TODO: Save volume name and uuid mapping in etcd as well
-            # This will make get_vol_byname more efficient
-            self._etcd.save_vol(vol)
-            LOG.debug('snapshot: %(name)s was successfully saved '
-                      'to etcd', {'name': snapshot_name})
-        except Exception as ex:
-            # TODO: 3PAR clean up issue over here - snapshot got
-            # created in the backend but since it could not be saved
-            # in ETCD db we are throwing an error saying operation
-            # failed.
-            # TODO: Snapshot needs clean up
-            msg = (_('save volume to etcd failed, error is: %s'),
-                   six.text_type(ex))
-            LOG.error(msg)
-            response = json.dumps({u"Err": six.text_type(ex)})
-        return response
+        return self._manager.create_snapshot(src_vol_name,
+                                             snapshot_name,
+                                             expiration_hrs,
+                                             retention_hrs)
 
     @app.route("/VolumeDriver.Mount", methods=["POST"])
     def volumedriver_mount(self, name):
@@ -830,75 +552,7 @@ class VolumePlugin(object):
             response = json.dumps({u"Err": "No Mount Point",
                                    u"Mountpoint": ""})
             return response
-
-        path_name = ''
-        path_info = self._etcd.get_vol_path_info(volname)
-
-        if path_info is not None:
-            path_name = path_info['mount_dir']
-
-        response = json.dumps({u"Err": '', u"Mountpoint": path_name})
-        return response
-
-    def _get_snapshots_to_be_deleted(self, db_snapshots, bkend_snapshots):
-        ss_list = []
-        for db_ss in db_snapshots:
-            found = False
-            bkend_ss_name = utils.get_3par_snap_name(db_ss['id'])
-
-            for bkend_ss in bkend_snapshots:
-                if bkend_ss_name == bkend_ss:
-                    found = True
-                    break
-            if not found:
-                ss_list.append(db_ss)
-        return ss_list
-
-    def _sync_snapshots_from_array(self, vol_id, db_snapshots):
-        bkend_snapshots = \
-            self.hpeplugin_driver.get_snapshots_by_vol(vol_id)
-        ss_list_remove = self._get_snapshots_to_be_deleted(db_snapshots,
-                                                           bkend_snapshots)
-        if ss_list_remove:
-            for ss in ss_list_remove:
-                db_snapshots.remove(ss)
-            self._etcd.update_vol(vol_id, 'snapshots',
-                                  db_snapshots)
-
-    def get_required_qos_field(self, qos_detail):
-        qos_filter = {}
-
-        msg = (_LI('get_required_qos_field: %(qos_detail)s'),
-               {'qos_detail': qos_detail})
-        LOG.info(msg)
-
-        qos_filter['enabled'] = qos_detail.get('enabled')
-
-        bwMaxLimitKB = qos_detail.get('bwMaxLimitKB')
-        if bwMaxLimitKB:
-            qos_filter['maxBWS'] = str(bwMaxLimitKB / 1024) + " MB/sec"
-
-        bwMinGoalKB = qos_detail.get('bwMinGoalKB')
-        if bwMinGoalKB:
-            qos_filter['minBWS'] = str(bwMinGoalKB / 1024) + " MB/sec"
-
-        ioMaxLimit = qos_detail.get('ioMaxLimit')
-        if ioMaxLimit:
-            qos_filter['maxIOPS'] = str(ioMaxLimit) + " IOs/sec"
-
-        ioMinGoal = qos_detail.get('ioMinGoal')
-        if ioMinGoal:
-            qos_filter['minIOPS'] = str(ioMinGoal) + " IOs/sec"
-
-        latencyGoal = qos_detail.get('latencyGoal')
-        if latencyGoal:
-            qos_filter['Latency'] = str(latencyGoal) + " sec"
-
-        priority = qos_detail.get('priority')
-        if priority:
-            qos_filter['priority'] = volume.QOS_PRIORITY[priority]
-
-        return qos_filter
+        return self._manager.get_path(volname)
 
     @app.route("/VolumeDriver.Get", methods=["POST"])
     def volumedriver_get(self, name):
@@ -909,17 +563,14 @@ class VolumePlugin(object):
 
         :return: Result indicating success.
         """
-        err = ''
-        mountdir = ''
-        devicename = ''
         contents = json.loads(name.content.getvalue())
-        volname = contents['Name']
-        tokens = volname.split('/')
+        qualified_name = contents['Name']
+        tokens = qualified_name.split('/')
         token_cnt = len(tokens)
 
         if token_cnt > 2:
             msg = (_LE("invalid volume or snapshot name %s"
-                       % volname))
+                       % qualified_name))
             LOG.error(msg)
             response = json.dumps({u"Err": msg})
             return response
@@ -929,79 +580,8 @@ class VolumePlugin(object):
         if token_cnt == 2:
             snapname = tokens[1]
 
-        volinfo = self._etcd.get_vol_byname(volname)
-
-        if volinfo is None:
-            msg = (_LE('Volume Get: Volume name not found %s'), volname)
-            LOG.warning(msg)
-            response = json.dumps({u"Err": ""})
-            return response
-
-        path_info = self._etcd.get_vol_path_info(volname)
-        if path_info is not None:
-            mountdir = path_info['mount_dir']
-            devicename = path_info['path']
-
-        # use volinfo as volname could be partial match
-        volume = {'Name': contents['Name'],
-                  'Mountpoint': mountdir,
-                  'Devicename': devicename,
-                  'Status': {}}
-
-        if volinfo['snapshots']:
-            self._sync_snapshots_from_array(volinfo['id'],
-                                            volinfo['snapshots'])
-        # Is this request for snapshot inspect?
-        if snapname:
-            # Any snapshots left after synchronization with array?
-            if volinfo['snapshots']:
-                snapshot, idx = \
-                    self._get_snapshot_by_name(volinfo['snapshots'],
-                                               snapname)
-                settings = {"Settings": {
-                    'expirationHours': snapshot['expiration_hours'],
-                    'retentionHours': snapshot['retention_hours']}}
-                volume['Status'].update(settings)
-            else:
-                msg = (_LE('Snapshot Get: Snapshot name not found %s'),
-                       contents['Name'])
-                LOG.warning(msg)
-                # Should error be returned here or success?
-                response = json.dumps({u"Err": ""})
-                return response
-
-        else:
-            snapshots = volinfo.get('snapshots', None)
-            if snapshots:
-                ss_list_to_show = []
-                for s in snapshots:
-                    snapshot = {'Name': s['name'],
-                                'ParentName': volname}
-                    ss_list_to_show.append(snapshot)
-                volume['Status'].update({'Snapshots': ss_list_to_show})
-
-        qos_name = volinfo.get('qos_name')
-        if qos_name is not None:
-            try:
-                qos_detail = self.hpeplugin_driver.get_qos_detail(qos_name)
-                qos_filter = self.get_required_qos_field(qos_detail)
-                volume['Status'].update({'qos_detail': qos_filter})
-            except Exception as ex:
-                msg = (_('unable to get/filter qos from 3par, error is: %s'),
-                       six.text_type(ex))
-                LOG.error(msg)
-                return json.dumps({u"Err": six.text_type(ex)})
-
-        vol_detail = {}
-        vol_detail['size'] = volinfo.get('size')
-        vol_detail['flash_cache'] = volinfo.get('flash_cache')
-        vol_detail['compression'] = volinfo.get('compression')
-        vol_detail['provisioning'] = volinfo.get('provisioning')
-        volume['Status'].update({'volume_detail': vol_detail})
-
-        response = json.dumps({u"Err": err, u"Volume": volume})
-        LOG.debug("Get volume/snapshot: \n%s" % str(response))
-        return response
+        return self._manager.get_volume_snap_details(
+            volname, snapname, qualified_name)
 
     @app.route("/VolumeDriver.List", methods=["POST"])
     def volumedriver_list(self, body):
@@ -1012,32 +592,7 @@ class VolumePlugin(object):
 
         :return: Result indicating success.
         """
-        volumes = self._etcd.get_all_vols()
-
-        if volumes is None:
-            response = json.dumps({u"Err": ''})
-            return response
-
-        volumelist = []
-        for volinfo in volumes.children:
-            if volinfo.key != util.VOLUMEROOT:
-                path_info = self._etcd.get_path_info_from_vol(volinfo.value)
-                if path_info is not None and 'mount_dir' in path_info:
-                    mountdir = path_info['mount_dir']
-                    devicename = path_info['path']
-                else:
-                    mountdir = ''
-                    devicename = ''
-                info = json.loads(volinfo.value)
-                volume = {'Name': info['display_name'],
-                          'Devicename': devicename,
-                          'size': info['size'],
-                          'Mountpoint': mountdir,
-                          'Status': {}}
-                volumelist.append(volume)
-
-        response = json.dumps({u"Err": '', u"Volumes": volumelist})
-        return response
+        return self._manager.list_volumes()
 
     def revert_to_snapshot(self, name, opts=None):
         contents = json.loads(name.content.getvalue())
@@ -1047,33 +602,4 @@ class VolumePlugin(object):
             raise exception.HPEPluginCreateException(reason=msg)
         snapname = contents['Name']
         volumename = str(contents['Opts']['promote'])
-        return self._revert_to_snapshot(volumename, snapname)
-
-    @synchronization.synchronized('{volumename}')
-    def _revert_to_snapshot(self, volumename, snapname):
-        volume = self._etcd.get_vol_byname(volumename)
-        if volume is None:
-            msg = (_LE('Volume: %s does not exist' % volumename))
-            LOG.info(msg)
-            response = json.dumps({u"Err": msg})
-            return response
-
-        snapshots = volume['snapshots']
-        LOG.info("Getting snapshot by name: %s" % snapname)
-        snapshot, idx = self._get_snapshot_by_name(snapshots, snapname)
-        if snapshot:
-            try:
-                LOG.info("Found snapshot by name %s" % snapname)
-                self.hpeplugin_driver.revert_snap_to_vol(volume, snapshot)
-                response = json.dumps({u"Err": ''})
-                return response
-            except Exception as ex:
-                msg = (_('revert snapshot failed, error is: %s'),
-                       six.text_type(ex))
-                LOG.error(msg)
-                return json.dumps({u"Err": six.text_type(ex)})
-        else:
-            msg = (_LE('snapshot: %s does not exist!' % snapname))
-            LOG.info(msg)
-            response = json.dumps({u"Err": msg})
-            return response
+        return self._manager.revert_to_snapshot(volumename, snapname)
