+++ test/createvolume_tester.py
@@ -1,5 +1,6 @@
 # import mock
 import fake_3par_data as data
+from hpedockerplugin import exception as hpe_exc
 import hpe_docker_unit_test as hpedockerunittest
 from hpe3parclient import exceptions
 from oslo_config import cfg
@@ -228,12 +229,10 @@ class TestCreateVolumeFlashCacheAddToVVSFails(CreateVolumeUnitTest):
         mock_3parclient.createVolumeSet.assert_called()
         mock_3parclient.modifyVolumeSet.assert_called()
         mock_3parclient.addVolumeToVolumeSet.assert_called()
-        mock_3parclient.deleteVolumeSet.assert_called()
 
-        # TODO: This is not happening at the moment and would make
-        # the unit test fail
-        # TODO: Remove comment once deleteVolume is invoked
-        # mock_3parclient.deleteVolume.assert_called()
+        # Rollback steps validation
+        mock_3parclient.deleteVolumeSet.assert_called()
+        mock_3parclient.deleteVolume.assert_called()
 
     def get_request_params(self):
         return {"Name": "test-vol-001",
@@ -339,6 +338,68 @@ class TestCreateCompressedVolNoHardwareSupport(CreateVolumeUnitTest):
         mock_3parclient.getStorageSystemInfo.return_value = \
             {'licenseInfo': {'licenses': []}}
 
+
+class TestCreateVolWithQosAndFlashCacheEtcdSaveFails(CreateVolumeUnitTest):
+    def check_response(self, resp):
+        expected = "ETCD data save failed: test-vol-001"
+        self._test_case.assertEqual(resp, {u"Err": expected})
+
+        mock_3parclient = self.mock_objects['mock_3parclient']
+        mock_3parclient.createVolume.assert_called()
+        mock_3parclient.modifyVolumeSet.assert_called()
+        mock_3parclient.addVolumeToVolumeSet.assert_called()
+
+        # Rollback steps validation
+        mock_3parclient.removeVolumeFromVolumeSet.assert_called()
+        mock_3parclient.deleteVolume.assert_called()
+
+    def get_request_params(self):
+        return {"Name": "test-vol-001",
+                "Opts": {"flash-cache": "true",
+                         "qos-name": "vvk_vvset",
+                         "provisioning": "thin",
+                         "size": "2"}}
+
+    def setup_mock_objects(self):
+        mock_etcd = self.mock_objects['mock_etcd']
+        mock_etcd.get_vol_byname.return_value = None
+        mock_etcd.save_vol.side_effect = \
+            [hpe_exc.HPEPluginSaveFailed(obj='test-vol-001')]
+
+        mock_3parclient = self.mock_objects['mock_3parclient']
+        mock_3parclient.getCPG.return_value = {}
+
+
+class TestCreateVolWithFlashCacheEtcdSaveFails(CreateVolumeUnitTest):
+    def check_response(self, resp):
+        expected = "ETCD data save failed: test-vol-001"
+        self._test_case.assertEqual(resp, {u"Err": expected})
+
+        mock_3parclient = self.mock_objects['mock_3parclient']
+        mock_3parclient.createVolume.assert_called()
+        mock_3parclient.createVolumeSet.assert_called()
+        mock_3parclient.modifyVolumeSet.assert_called()
+        mock_3parclient.addVolumeToVolumeSet.assert_called()
+
+        # Rollback steps validation
+        mock_3parclient.removeVolumeFromVolumeSet.assert_called()
+        mock_3parclient.deleteVolumeSet.assert_called()
+        mock_3parclient.deleteVolume.assert_called()
+
+    def get_request_params(self):
+        return {"Name": "test-vol-001",
+                "Opts": {"flash-cache": "true",
+                         "provisioning": "thin",
+                         "size": "2"}}
+
+    def setup_mock_objects(self):
+        mock_etcd = self.mock_objects['mock_etcd']
+        mock_etcd.get_vol_byname.return_value = None
+        mock_etcd.save_vol.side_effect = \
+            [hpe_exc.HPEPluginSaveFailed(obj='test-vol-001')]
+
+        mock_3parclient = self.mock_objects['mock_3parclient']
+        mock_3parclient.getCPG.return_value = {}
 
 # More cases of flash cache
 # 1.
